utils: move git functions to pkg/gitutils
